Add variable to customize RLS executable
The suggested way of running the RLS is now through `rustup run nightly rls`.
Since this can change, there's no point of hardcoding it in lsp-rust.

This commit adds a customizable variable to use as the default command.  If the
variable is `nil`, lsp-rust will fallback to use the RLS defined by RLS_ROOT, if
any.